Handler для вебхука

+++ src/AmoApiClient/src/Handler/UnisWebhookHandler.php
@@ -6,13 +6,14 @@ namespace AmoApiClient\Handler;
 
 use DataBase\Services\ApiToken\get\GetApiTokenService;
 use DataBase\Services\Contact\create\Interfaces\SaveContactInterface;
+use DataBase\Services\User\get\Interfaces\GetUserInterface;
 use Laminas\Diactoros\Response\JsonResponse;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\RequestHandlerInterface;
-use UnisenderApi\Services\PrepareData\Interfaces\GetPreparedContactsForDeleteInterface;
-use UnisenderApi\Services\PrepareData\Interfaces\GetPreparedContactsForImportInterface;
-use UnisenderApi\Services\PrepareData\Interfaces\GetPreparedContactsForUpdateInterface;
+use UnisenderApi\Services\PrepareData\GetPreparedContactsForAddService;
+use UnisenderApi\Services\PrepareData\GetPreparedContactsForDeleteService;
+use UnisenderApi\Services\PrepareData\GetPreparedContactsForUpdateService;
 use UnisenderApi\Services\UnisenderApi\Interfaces\GetUnisenderApiInterface;
 
 /**
@@ -20,20 +21,22 @@ use UnisenderApi\Services\UnisenderApi\Interfaces\GetUnisenderApiInterface;
  */
 class UnisWebhookHandler implements RequestHandlerInterface
 {
-    private GetPreparedContactsForImportInterface $contactsForImport;
+    private GetPreparedContactsForAddService $contactsForImport;
     private GetApiTokenService $getApiToken;
     private GetUnisenderApiInterface $getUnisenderApi;
-    private GetPreparedContactsForDeleteInterface $contactsForDelete;
+    private GetPreparedContactsForDeleteService $contactsForDelete;
     private SaveContactInterface $saveContact;
-    private GetPreparedContactsForUpdateInterface $contactsForUpdate;
+    private GetPreparedContactsForUpdateService $contactsForUpdate;
+    private GetUserInterface $getUser;
 
     public function __construct(
-        GetPreparedContactsForImportInterface $contactsForImport,
-        GetPreparedContactsForDeleteInterface $contactsForDelete,
-        GetPreparedContactsForUpdateInterface $contactsForUpdate,
+        GetPreparedContactsForAddService $contactsForImport,
+        GetPreparedContactsForDeleteService $contactsForDelete,
+        GetPreparedContactsForUpdateService $contactsForUpdate,
         GetApiTokenService $getApiToken,
         GetUnisenderApiInterface $getUnisenderApi,
-        SaveContactInterface $saveContact
+        SaveContactInterface $saveContact,
+        GetUserInterface $getUser
     ) {
         $this->contactsForImport = $contactsForImport;
         $this->getApiToken = $getApiToken;
@@ -41,6 +44,7 @@ class UnisWebhookHandler implements RequestHandlerInterface
         $this->contactsForDelete = $contactsForDelete;
         $this->saveContact = $saveContact;
         $this->contactsForUpdate = $contactsForUpdate;
+        $this->getUser = $getUser;
     }
 
     public function handle(ServerRequestInterface $request): ResponseInterface
@@ -56,8 +60,15 @@ class UnisWebhookHandler implements RequestHandlerInterface
         if ($postParams['contacts']['add']) {
             $preparedContacts = $this->contactsForImport
                 ->prepare($postParams['contacts']['add'][0]);
+            $user = $this->getUser
+                ->get('account_id', (int) $postParams['account']['id']);
+
             $contact = $this->saveContact
-                ->save((int) $preparedContacts['bd']['contact_id']);
+                ->save(
+                    (int) $preparedContacts['bd']['contact_id'],
+                    $user->id,
+                );
+
             $contact->emails()
                 ->createMany($preparedContacts['bd']['emails']);
         }
@@ -65,17 +76,24 @@ class UnisWebhookHandler implements RequestHandlerInterface
         //При обновлении контактов
         if ($postParams['contacts']['update']) {
            $preparedContacts = $this->contactsForUpdate
-               ->prepare($postParams['contacts']['update'][0]);
+               ->prepare($postParams);
+           if ($preparedContacts === null) {
+               return new JsonResponse('Contacts updated', 200);
+           }
            $contact = $preparedContacts['bd']['contact_model'];
 
             // Добавляем новые email адреса
-            $contact->emails()
-                ->createMany($preparedContacts['bd']['emails']['create']);
+            if ($preparedContacts['bd']['emails']['create']) {
+                $contact->emails()
+                    ->createMany($preparedContacts['bd']['emails']['create']);
+            }
 
             // Удаляем отсутствующие email адреса
-            $emailsToDelete = $preparedContacts['bd']['emails']['delete'];
-            foreach ($emailsToDelete as $emailToDelete) {
-                $contact->emails()->where('email', $emailToDelete)->delete();
+            if ($preparedContacts['bd']['emails']['delete']) {
+                $emailsToDelete = $preparedContacts['bd']['emails']['delete'];
+                foreach ($emailsToDelete as $emailToDelete) {
+                    $contact->emails()->where('email', $emailToDelete)->delete();
+                }
             }
         }
 
@@ -83,14 +101,13 @@ class UnisWebhookHandler implements RequestHandlerInterface
         if ($postParams['contacts']['delete']) {
 
             $preparedContacts = $this->contactsForDelete
-                ->prepare($postParams['contacts']['delete'][0]);
+                ->prepare($postParams);
 
             $contact = $preparedContacts['bd']['contact_model'];
             $contact->delete();
         }
 
         $response = $unisenderApi->importContacts($preparedContacts['unis']);
-
 
         return new JsonResponse($response, 200);
     }
